vision_desktop_app/requirements.txt
tensorflow>=2.12
PySide6>=6.5
opencv-python
numpy
pillow
matplotlib
tqdm
Flask
